correct indentation for tabs, removed services folder.
