+++ back/src/app.module.ts
@@ -1,6 +1,6 @@
 import { Global, Module } from '@nestjs/common';
-import { Api42Service } from './services/api42.service';
-import { UsersService } from './services/users.service';
+import { Api42Service } from './modules/auth/api42.service';
+import { UsersService } from './modules/users/users.service';
 import { PrismaService } from './modules/prisma/prisma.service';
 import { AuthModule } from './modules/auth/auth.module';
 import { PrismaModule } from './modules/prisma/prisma.module';
@@ -9,9 +9,9 @@ import { CdnModule } from './modules/cdn/cdn.module';
 
 @Global()
 @Module({
-    imports: [AuthModule, PrismaModule, CdnModule, UsersModule],
-    controllers: [],
-    providers: [Api42Service, PrismaService, UsersService],
-    exports: [PrismaModule],
+	imports: [AuthModule, PrismaModule, CdnModule, UsersModule],
+	controllers: [],
+	providers: [Api42Service, PrismaService, UsersService],
+	exports: [PrismaModule],
 })
 export class AppModule {}
