+++ back/src/modules/auth/DTO/RefreshTokenDTO.ts
@@ -1,7 +1,7 @@
 import { IsString, MinLength } from 'class-validator';
 
 export class RefreshTokenDTO {
-    @IsString()
-    @MinLength(1)
-    refresh_token: string;
+	@IsString()
+	@MinLength(1)
+	refresh_token: string;
 }
