+++ back/src/modules/cdn/cdn.controller.ts
@@ -1,13 +1,13 @@
 import { Controller, Get, Param, Res } from '@nestjs/common';
-import { UsersService } from 'src/services/users.service';
+import { UsersService } from '../users/users.service';
 
 @Controller('cdn')
 export class CdnController {
-    constructor(private readonly userService: UsersService) {}
+	constructor(private readonly userService: UsersService) {}
 
-    @Get('user/:userLogin')
-    async user(@Param('userLogin') userLogin: string, @Res() res) {
-        const imgPath = `./cdn/profile_pictures/${userLogin}`;
-        return res.sendFile(imgPath, { root: 'public' });
-    }
+	@Get('user/:userLogin')
+	async user(@Param('userLogin') userLogin: string, @Res() res) {
+		const imgPath = `./cdn/profile_pictures/${userLogin}`;
+		return res.sendFile(imgPath, { root: 'public' });
+	}
 }
