+++ back/src/modules/cdn/cdn.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { UsersService } from 'src/services/users.service';
+import { UsersService } from '../users/users.service';
 import { CdnController } from './cdn.controller';
 
 @Module({
-    imports: [],
-    controllers: [CdnController],
-    providers: [UsersService],
+	imports: [],
+	controllers: [CdnController],
+	providers: [UsersService],
 })
 export class CdnModule {}
