+++ back/src/modules/prisma/prisma.service.ts
@@ -3,18 +3,18 @@ import { PrismaClient } from '@prisma/client';
 
 @Injectable()
 export class PrismaService
-    extends PrismaClient
-    implements OnModuleInit, OnModuleDestroy
+extends PrismaClient
+implements OnModuleInit, OnModuleDestroy
 {
-    constructor() {
-        super();
-    }
+	constructor() {
+		super();
+	}
 
-    async onModuleInit() {
-        await this.$connect();
-    }
+	async onModuleInit() {
+		await this.$connect();
+	}
 
-    async onModuleDestroy() {
-        await this.$disconnect();
-    }
+	async onModuleDestroy() {
+		await this.$disconnect();
+	}
 }
