+++ back/src/modules/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
-import { UsersService } from 'src/services/users.service';
+import { UsersService } from './users.service';
 
 @Module({
 	imports: [],
